requirements.txt
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
